+++ test/support/mysql/migrations/02_create_lotus_tables.exs
@@ -1,0 +1,6 @@
+defmodule Lotus.Test.MysqlRepo.Migrations.CreateLotusTables do
+  use Ecto.Migration
+
+  defdelegate up, to: Lotus.Migrations.MySQL
+  defdelegate down, to: Lotus.Migrations.MySQL
+end